Ensure that destroyed cities no longer show up in the UI
Now if a city is captured it is removed from the UI and the implicit
city road is removed.

To make this easier to test I movified `c7-static-map-save.json` locally
to add this:

```
	{
		"id": "Warrior-1",
		"prototype": "Warrior",
		"owner": "player-2",
		"previousLocation": {
		"x": -1,
		"y": -1
		},
		"currentLocation": {
		"x": 57,
		"y": 43
		},
		"hitPointsRemaining": 5,
		"movePointsRemaining": 1,
		"facingDirection": "north",
		"experience": "Elite"
	},
```

and modified this (to change the x/y position so an AI spawns next to
us)

```
	{
		"id": "Settler-2",
		"prototype": "Settler",
		"owner": "player-3",
		"previousLocation": {
		"x": -1,
		"y": -1
		},
		"currentLocation": {
		"x": 59,
		"y": 43
		},
		"hitPointsRemaining": 3,
		"movePointsRemaining": 1,
		"facingDirection": "north",
		"experience": "Regular"
	},
```

I also removed some unused variables that I noticed.